+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0
+
+- Fix `seeds_path` to be relative to Rails root
+- Improve test coverage for generator
+
 # 1.0.1
 
 - Fix boolean config options to properly handle `nil` and `false` values